requirements.txt
absl-py==0.11.0
appdirs==1.4.4
apturl==0.5.2
astroid==2.4.2
astunparse==1.6.3
autopep8==1.5.5
bandit==1.7.0
black==20.8b1
blinker==1.4
Brlapi==0.7.0
cachetools==4.2.1
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
docopt==0.6.2
EasyProcess==0.3
entrypoint2==0.2.3
entrypoints==0.3
evdev==1.4.0
flatbuffers==1.12
gast==0.3.3
gitdb==4.0.5
GitPython==3.1.13
google-auth==1.26.1
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
httplib2==0.14.0
idna==2.8
isort==5.7.0
jeepney==0.6.0
joblib==1.0.1
Keras==2.4.3
Keras-Preprocessing==1.1.2
keyboard==0.13.5
keyring==18.0.1
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.3
MarkupSafe==1.1.0
matplotlib==3.3.4
mccabe==0.6.1
MouseInfo==0.1.3
mss==6.1.0
mypy-extensions==0.4.3
netifaces==0.10.4
numpy==1.19.5
oauthlib==3.1.0
olefile==0.46
opencv-python==4.5.1.48
opt-einsum==3.3.0
packaging==20.3
pandas==1.2.2
pathspec==0.8.1
pbr==5.5.1
pexpect==4.6.0
Pillow==7.0.0
pipreqs==0.4.10
protobuf==3.14.0
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.52
pycairo==1.16.2
pycodestyle==2.6.0
pycups==1.9.73
pygame==2.0.1
PyGetWindow==0.0.9
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pylint==2.6.0
pymacaroons==0.13.0
PyMsgBox==1.0.9
PyNaCl==1.3.0
pynput==1.7.3
pyparsing==2.4.6
pyperclip==1.8.1
PyRect==0.1.4
pyRFC3339==1.1
pyscreenshot==2.3
PyScreeze==0.1.26
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-xlib==0.29
python3-xlib==0.15
PyTweening==1.0.3
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
regex==2020.11.13
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.7
scikit-learn==0.24.1
scipy==1.6.0
screen-resolution-extra==0.0.0
seaborn==0.11.1
SecretStorage==2.3.1
simplejson==3.16.0
six==1.15.0
sklearn==0.0
smmap==3.0.5
stevedore==3.3.0
systemd-python==234
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
tflearn==0.5.0
threadpoolctl==2.1.0
toml==0.10.2
torch==1.7.1+cu101
torchaudio==0.7.2
torchvision==0.8.2+cu101
tqdm==4.56.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
wadllib==1.3.3
Werkzeug==1.0.1
wrapt==1.12.1
xkit==0.0.0
yarg==0.1.9
